Change of structure and iteration between screens
