+++ src/lib/types/viewType.ts
@@ -1,0 +1,6 @@
+export const VIEWS = [
+    "grid", "list", "fullgrid", "favorites"] as const;
+
+export type ViewType = 
+    typeof VIEWS[number];
+
